+++ src/DependencyInjection/Configuration.php
@@ -18,6 +18,9 @@ namespace Hofff\Contao\RateIt\DependencyInjection;
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
+use function func;
+use function is_array;
+use function var_dump;
 
 final class Configuration implements ConfigurationInterface
 {
@@ -29,23 +32,54 @@ final class Configuration implements ConfigurationInterface
         $rootNode
             ->children()
                 ->arrayNode('types')
+                    ->info('Enable or disable ratings for different content types')
                     ->addDefaultsIfNotSet()
                     ->children()
                         ->booleanNode('page')
+                            ->info('If enabled rating for pages can be configured')
                             ->defaultTrue()
                         ->end()
                         ->booleanNode('article')
+                            ->info('If enabled rating for articles can be configured')
                             ->defaultTrue()
                         ->end()
                         ->booleanNode('news')
+                            ->info('If enabled rating for new can be configured')
                             ->defaultTrue()
                         ->end()
                         ->booleanNode('module')
+                            ->info('If enabled rating module can be configured')
                             ->defaultTrue()
                         ->end()
                         ->booleanNode('ce')
+                            ->info('If enabled rating content element can be configured')
                             ->defaultTrue()
                         ->end()
+                        ->booleanNode('comments')
+                            ->info('If enabled rating for comments can be configured')
+                            ->defaultTrue()
+                        ->end()
+                    ->end()
+                ->end()
+                ->arrayNode('comment_sources')
+                    ->info('Defines in which source the rating is activated for comments')
+                    ->scalarPrototype()
+                    ->end()
+                    ->beforeNormalization()
+                        ->always()
+                        ->then(
+                            static function ($value) {
+                                if (!is_array($value)) {
+                                    return ['tl_news' => 'tl_news_archive'];
+                                }
+
+                                if (!isset($value['tl_news'])) {
+                                    $value['tl_news'] = 'tl_news_archive';
+                                }
+
+                                return $value;
+                            }
+                        )
                     ->end()
                 ->end()
             ->end();